splitting aot tests by module format
